Refactor datawrapper
