--- src/model/openthaigpt_pretraining_model/GPTJ_TH_tokenizer/__init__ .py
@@ -1,2 +1,0 @@
-"""Top-level package for OpenThaiGPT Model."""
-__version__ = "0.1.0"
