+++ src/model/openthaigpt_pretraining_model/lightning/constants.py
@@ -1,6 +1,6 @@
 DATASET_NAME = "c4"
 LANGUAGE_DATASET = "en"
-SPLIT_TRAIN = "train"
 SPLIT_VAL = "validation"
+SPLIT_TRAIN = "train"
 LLAMA_MODEL = "decapoda-research/llama-7b-hf"
 GPTJ_MODEL = "EleutherAI/gpt-j-6B"
