+++ src/model/openthaigpt_pretraining_model/lightning/utils.py
@@ -1,13 +1,11 @@
 from datasets import load_dataset
-import numpy as np
-import random
 from tqdm import tqdm
 import lightning as L
 from lightning.fabric.accelerators import Accelerator
 from lightning.fabric.strategies import Strategy
 import torch
 import torch.optim as optim
-from torch.utils.data import IterableDataset, DataLoader
+from torch.utils.data import DataLoader
 from lion_pytorch import Lion
 from typing import List, Union
 from transformers import (
@@ -29,6 +27,8 @@ from openthaigpt_pretraining_model.models.llama.model import make_model_llama
 from openthaigpt_pretraining_model.models.llama_hf.model import (
     make_model_llama_hf,
 )
+from ..utils import compute_perplexity
+from ..data_wrapper import DatasetWrapper
 from lightning.fabric.strategies import DeepSpeedStrategy
 import deepspeed
 import wandb
@@ -36,57 +36,6 @@ import os
 
 # os.environ["WANDB_API_KEY"] = "<your-api-key>"
 os.environ["WANDB_MODE"] = "offline"
-
-
-class DatasetWrapper(IterableDataset):
-    def __init__(self, mode, model, max_tokens=256):
-        if model != "decapoda-research/llama-7b-hf":
-            self.tokenizer = AutoTokenizer.from_pretrained(model)
-        else:
-            self.tokenizer = LlamaTokenizer.from_pretrained(model)
-        self.mode = mode
-        self.max_tokens = max_tokens
-
-        if mode == "val":
-            self.data_set = load_dataset(
-                DATASET_NAME,
-                LANGUAGE_DATASET,
-                streaming=True,
-                split=SPLIT_VAL,
-            )
-        elif mode == "train":
-            self.data_set = load_dataset(
-                DATASET_NAME,
-                LANGUAGE_DATASET,
-                streaming=True,
-                split=SPLIT_TRAIN,
-            ).shuffle(buffer_size=10_000)
-        else:
-            raise NotImplementedError("only support Train,Val")
-
-    def __iter__(self):
-        buffer = []
-        iter_dataset = self.data_set
-
-        for sample in iter_dataset:
-            buffer += self.tokenizer(sample["text"])["input_ids"]
-            while len(buffer) > self.max_tokens:
-                yield torch.tensor(buffer[: self.max_tokens])
-                buffer = buffer[self.max_tokens :]
-
-
-def seed_everything(seed):
-    torch.manual_seed(seed)
-    torch.cuda.manual_seed(seed)
-    torch.cuda.manual_seed_all(seed)
-    torch.backends.cudnn.deterministic = True
-    torch.backends.cudnn.benchmark = False
-    np.random.seed(seed)
-    random.seed(seed)
-
-
-def compute_perplexity(loss: torch.Tensor) -> float:
-    return torch.exp(loss).item()
 
 
 class Trainer:
@@ -144,6 +93,7 @@ class Trainer:
 
         if model_name == "llama":
             model_name = LLAMA_MODEL  # for tokenizer
+            self.tokenizer = LlamaTokenizer.from_pretrained(model_name)
             self.model = make_model_llama(
                 vocab_size=vocab_size,
                 context_length=context_length,
@@ -153,6 +103,7 @@ class Trainer:
             )
         elif model_name == "llama_hf":
             model_name = LLAMA_MODEL  # for tokenizer
+            self.tokenizer = LlamaTokenizer.from_pretrained(model_name)
             self.model = make_model_llama_hf(
                 vocab_size=vocab_size,
                 context_length=context_length,
@@ -161,6 +112,7 @@ class Trainer:
             )
         elif model_name == "gptj":
             model_name = GPTJ_MODEL  # for tokenizer
+            self.tokenizer = AutoTokenizer.from_pretrained(model_name)
             self.model = make_model_gptj(
                 vocab_size=vocab_size,
                 context_length=context_length,
@@ -171,9 +123,20 @@ class Trainer:
             )
         else:
             raise NotImplementedError("only support Llama, llama_hf or GPTJ")
-
-        self.dataset = DatasetWrapper("train", model_name, self.max_tokens)
-        self.dataset_val = DatasetWrapper("val", model_name, self.max_tokens)
+        train_dataset = load_dataset(
+            DATASET_NAME,
+            LANGUAGE_DATASET,
+            streaming=True,
+            split=SPLIT_TRAIN,
+        ).shuffle(buffer_size=10_000)
+        val_dataset = load_dataset(
+            DATASET_NAME,
+            LANGUAGE_DATASET,
+            streaming=True,
+            split=SPLIT_VAL,
+        )
+        self.dataset = DatasetWrapper(self.tokenizer, train_dataset, self.max_tokens)
+        self.dataset_val = DatasetWrapper(self.tokenizer, val_dataset, self.max_tokens)
         self.tokenizer = self.dataset.tokenizer
         self.dataloader = DataLoader(
             self.dataset,
