+++ src/model/openthaigpt_pretraining_model/optimizers/lion/utils.py
@@ -5,12 +5,10 @@ import os
 import torch
 
 import torch.optim as optim
-from torch.utils.data import DataLoader, IterableDataset
+from torch.utils.data import DataLoader
 from torch.nn.parallel import DistributedDataParallel as DDP  # noqa: N817
 from torch.distributed import init_process_group, destroy_process_group
 
-import numpy as np
-import random
 from tqdm import tqdm
 
 from datasets import load_dataset
@@ -33,18 +31,9 @@ from .constants import (
     SPLIT_TRAIN,
     LANGUAGE_DATASET,
 )
+from ...data_wrapper import DatasetWrapper
 
 _attn_orig = GPT2Attention._attn
-
-
-def seed_everything(seed):
-    torch.manual_seed(seed)
-    torch.cuda.manual_seed(seed)
-    torch.cuda.manual_seed_all(seed)
-    torch.backends.cudnn.deterministic = True
-    torch.backends.cudnn.benchmark = False
-    np.random.seed(seed)
-    random.seed(seed)
 
 
 def closest_power_of_2(x):
@@ -84,45 +73,6 @@ def get_torch_context(dtype: str):
     return ctx
 
 
-class DatasetWrapper(IterableDataset):
-    def __init__(self, mode, max_tokens=256):
-        self.tokenizer = GPT2TokenizerFast.from_pretrained(
-            MODEL_NAME,
-            bos_token=BOS_TOKEN,
-            eos_token=EOS_TOKEN,
-            pad_token=PAD_TOKEN,
-        )
-        self.mode = mode
-        self.max_tokens = max_tokens
-
-        if mode == "val":
-            self.data_set = load_dataset(
-                DATASET_NAME,
-                languages=[LANGUAGE_DATASET],
-                streaming=True,
-                split=SPLIT_VAL,  # optional
-            )
-        elif mode == "train":
-            self.data_set = load_dataset(
-                DATASET_NAME,
-                languages=[LANGUAGE_DATASET],
-                streaming=True,
-                split=SPLIT_TRAIN,  # optional
-            ).shuffle(buffer_size=10_000)
-        else:
-            raise NotImplementedError("only support Train,Val")
-
-    def __iter__(self):
-        buffer = []
-        iter_dataset = self.data_set
-
-        for sample in iter_dataset:
-            buffer += self.tokenizer(sample["text"] + EOS_TOKEN)["input_ids"]
-            while len(buffer) > self.max_tokens:
-                yield torch.tensor(buffer[: self.max_tokens])
-                buffer = buffer[self.max_tokens :]
-
-
 class Trainer:
     def __init__(
         self,
@@ -151,8 +101,26 @@ class Trainer:
         self.warmup_steps = warmup_steps
         self.eval_steps = eval_steps
         self.do_sample = do_sample
-        self.dataset = DatasetWrapper("train", self.max_tokens)
-        self.dataset_val = DatasetWrapper("val", self.max_tokens)
+        tokenizer = GPT2TokenizerFast.from_pretrained(
+            MODEL_NAME,
+            bos_token=BOS_TOKEN,
+            eos_token=EOS_TOKEN,
+            pad_token=PAD_TOKEN,
+        )
+        dataset_train = load_dataset(
+            DATASET_NAME,
+            languages=[LANGUAGE_DATASET],
+            streaming=True,
+            split=SPLIT_TRAIN,  # optional
+        ).shuffle(buffer_size=10_000)
+        dataset_val = load_dataset(
+            DATASET_NAME,
+            languages=[LANGUAGE_DATASET],
+            streaming=True,
+            split=SPLIT_VAL,  # optional
+        )
+        self.dataset = DatasetWrapper(tokenizer, dataset_train, self.max_tokens)
+        self.dataset_val = DatasetWrapper(tokenizer, dataset_val, self.max_tokens)
         self.use_flash = use_flash
         self.use_checkpointing = use_checkpointing
         self.use_rotary = use_rotary
