+++ src/model/openthaigpt_pretraining_model/utils.py
@@ -1,0 +1,17 @@
+import random
+import numpy as np
+import torch
+
+
+def seed_everything(seed):
+    torch.manual_seed(seed)
+    torch.cuda.manual_seed(seed)
+    torch.cuda.manual_seed_all(seed)
+    torch.backends.cudnn.deterministic = True
+    torch.backends.cudnn.benchmark = False
+    np.random.seed(seed)
+    random.seed(seed)
+
+
+def compute_perplexity(loss: torch.Tensor) -> float:
+    return torch.exp(loss).item()
